Unit-Prefixed Bytes

Index

PC2GPUBenchmarks.bytes

PC2GPUBenchmarks.bytes

PC2GPUBenchmarks.change_base

PC2GPUBenchmarks.simplify

PC2GPUBenchmarks.B

PC2GPUBenchmarks.GB

PC2GPUBenchmarks.GiB

PC2GPUBenchmarks.KB

PC2GPUBenchmarks.KiB

PC2GPUBenchmarks.MB

PC2GPUBenchmarks.MiB

PC2GPUBenchmarks.TB

PC2GPUBenchmarks.TiB

PC2GPUBenchmarks.UnitPrefixedBytes

References

PC2GPUBenchmarks.B — Type

Bytes

PC2GPUBenchmarks.GB — Type

Gigabytes, i.e. 10^9 = 1000^3 bytes

PC2GPUBenchmarks.GiB — Type

Gibibytes, i.e. 2^30 = 1024^3 bytes

PC2GPUBenchmarks.KB — Type

Kilobytes, i.e. 10^3 = 1000 bytes

PC2GPUBenchmarks.KiB — Type

Kibibytes, i.e. 2^10 = 1024 bytes

PC2GPUBenchmarks.MB — Type

Megabytes, i.e. 10^6 = 1000^2 bytes

PC2GPUBenchmarks.MiB — Type

Mebibytes, i.e. 2^20 = 1024^2 bytes

PC2GPUBenchmarks.TB — Type

Terabytes, i.e. 10^12 = 1000^4 bytes

PC2GPUBenchmarks.TiB — Type

Tebibytes, i.e. 2^40 = 1024^4 bytes

PC2GPUBenchmarks.UnitPrefixedBytes — Type

Abstract type representing an amount of data, i.e. a certain number of bytes, with a unit prefix (also "metric prefix"). Examples include the SI prefixes, like KB, MB, and GB, but also the binary prefixes (ISO/IEC 80000), like KiB, MiB, and GiB.

See https://en.wikipedia.org/wiki/Binary_prefix for more information.

PC2GPUBenchmarks.bytes — Method

bytes(x::Number)

Returns an appropriate UnitPrefixedBytes object, representing the number of bytes.

Note: This function is type unstable by construction!

See simplify for what "appropriate" means here.

PC2GPUBenchmarks.bytes — Method

bytes(x::UnitPrefixedBytes)

Return the number of bytes (without prefix) as Float64.

PC2GPUBenchmarks.change_base — Method

Toggle between

Base 10, SI prefixes, i.e. factors of 1000

Base 2, ISO/IEC prefixes, i.e. factors of 1024

Example:

julia> change_base(KB(13))
~12.7 KiB

julia> change_base(KiB(13))
~13.31 KB

PC2GPUBenchmarks.simplify — Method

simplify(x::UnitPrefixedBytes[; base])

Given a UnitPrefixedBytes number x, finds a more appropriate UnitPrefixedBytes that represents the same number of bytes but with a smaller value.

The optional keyword argument base can be used to switch between base 2, i.e. ISO/IEC prefixes (default), and base 10. Allowed values are 2, 10, :SI, :ISO, and :IEC.

Note: This function is type unstable by construction!

Example:

julia> simplify(B(40_000_000))
~38.15 MiB

julia> simplify(B(40_000_000); base=10)
40.0 MB