GPU Stresstest

Index

PC2GPUBenchmarks.stresstest

PC2GPUBenchmarks.StressTest

PC2GPUBenchmarks.StressTestStoreResults

References

PC2GPUBenchmarks.stresstest — Method

stresstest(device_or_devices)

Run a GPU stress test (matrix multiplication) on one or multiple GPU devices, as specified by the positional argument. If no argument is provided (only) the currently active GPU will be used.

Keyword arguments:

duration (default: nothing): desired run time in seconds. Takes precedence over mem if also set.

mem (default: nothing): number (<:Real) between 0 and 1, indicating the fraction of the available GPU memory that should be used, or a <:UnitPrefixedBytes indicating an absolute memory limit. May not be used together with duration.

dtype (default: Float32): element type of the matrices

size (default: 2048): matrices of size (size, size) will be used

verbose (default: true): toggle printing of information

parallel (default: true): If true, will (try to) run each GPU test on a different Julia thread. Make sure to have enough Julia threads.

threads (default: nothing): If parallel == true, this argument may be used to specify the Julia threads to use.

PC2GPUBenchmarks.StressTest — Type

GPU stress test (matrix multiplications) in which we try to run for a given time period.

PC2GPUBenchmarks.StressTestStoreResults — Type

GPU stress test (matrix multiplications) in which we store all matmul results and try to run as many iterations as possible for a certain memory limit (default: 90% of free memory).

This stress test is somewhat inspired by gpu-burn by Ville Timonen.