GPU Stresstest

Somewhat inspired by gpu-burn.

Index

PC2GPUBenchmarks.gputest

PC2GPUBenchmarks.GPUTest_FixedDuration

PC2GPUBenchmarks.GPUTest_StoreResults

References

PC2GPUBenchmarks.GPUTest_FixedDuration — Type

GPU stress test (matrix multiplications) in which we try to run for a given time period.

PC2GPUBenchmarks.GPUTest_StoreResults — Type

GPU stress test (matrix multiplications) in which we store all matmul results and try to run as many iterations as possible for a certain memory limit (default: 90% of free memory).

PC2GPUBenchmarks.gputest — Function

gputest(device_or_devices)

Run a GPU Test (matrix multiplication) on one or multiple GPU devices, as specified by the positional argument. If no argument is provided (only) the currently active GPU will be used.

Keyword arguments:

duration (default: nothing): desired run time in seconds. Takes precedence over mem if also set.

mem (default: nothing): number (<:Real) between 0 and 1, indicating the fraction of the available GPU memory that should be used, or a <:UnitPrefixedBytes indicating an absolute memory limit. May not be used together with duration.

dtype (default: Float32): element type of the matrices

size (default: 2048): matrices of size (size, size) will be used

verbose (default: true): toggle printing of information

If multiple devices are provided there are additional keyword arguments:

parallel (default: true): If true, will (try to) run each GPU test in on a different Julia thread. Make sure to have enough Julia threads.

threads (default: nothing): If parallel == true, this argument may be used to specify the Julia threads to use.