Memory Bandwidth (SAXPY GPU)

Index

PC2GPUBenchmarks.saxpy_gpu

PC2GPUBenchmarks.saxpy_gpu_scaling

References

PC2GPUBenchmarks.saxpy_gpu — Method

Tries to estimate the peak memory bandwidth of a GPU in GiB/s by measuring the time it takes to perform a SAXPY, i.e. a * x[i] + y[i].

Keyword arguments:

device (default: CUDA.device()): CUDA device to be used.

dtype (default: Float32): element type of the vectors.

size (default: 2^20 * 10): length of the vectors.

nbench (default: 5): number of measurements to be performed the best of which is used for the GiB/s computation.

verbose (default: true): toggle printing.

cublas (default: true): toggle between CUDA.axpy! and a custom saxpy_gpu_kernel!.

See also: saxpy_gpu_scaling.

PC2GPUBenchmarks.saxpy_gpu_scaling — Method

saxpy_gpu_scaling() -> sizes, bandwidths

Measures the memory bandwidth (via saxpy_gpu) as a function of vector length. If verbose=true (default), displays a unicode plot. Returns the considered lengths and GiB/s. For further options, see saxpy_gpu.