Utility

Index

PC2GPUBenchmarks.alloc_mem

PC2GPUBenchmarks.clear_all_gpus_memory

PC2GPUBenchmarks.clear_gpu_memory

PC2GPUBenchmarks.functional

PC2GPUBenchmarks.get_power_usage

PC2GPUBenchmarks.get_temperature

PC2GPUBenchmarks.gpuid

PC2GPUBenchmarks.monitor_powerusage

PC2GPUBenchmarks.monitor_something

PC2GPUBenchmarks.monitor_temperature

References

PC2GPUBenchmarks.alloc_mem — Method

alloc_mem(memsize::UnitPrefixedBytes; devs=(CUDA.device(),), dtype=Float32)

Allocates memory on the devices whose IDs are provided via devs. Returns a vector of memory handles (i.e. CuArrays).

Examples:

alloc_mem(MiB(1024)) # allocate on the currently active device
alloc_mem(B(40_000_000); devs=(0,1)) # allocate on GPU0 and GPU1

PC2GPUBenchmarks.clear_all_gpus_memory — Function

Reclaim the unused memory of all available GPUs.

PC2GPUBenchmarks.clear_gpu_memory — Function

Reclaim the unused memory of the currently active GPU (i.e. device()).

PC2GPUBenchmarks.functional — Function

Check if CUDA/GPU is available and functional. If not, print some (hopefully useful) debug information.

PC2GPUBenchmarks.get_power_usage — Method

Get current power usage of given device in Watts.

PC2GPUBenchmarks.get_temperature — Method

Get current temperature of given device in degrees Celsius.

PC2GPUBenchmarks.gpuid — Method

Get GPU index of the given device.

Note: GPU indices start with zero.

PC2GPUBenchmarks.monitor_powerusage — Method

monitor_powerusage(duration) -> times, powerusage

Monitor the power usage of GPU(s) (in Watts) over a given time period, as specified by duration (in seconds). Returns the (relative) times as a Vector{Float64} and the power usage as a Vector{Vector{Float64}}.

For general keyword arguments, see monitor_something.

PC2GPUBenchmarks.monitor_something — Method

monitor_something(f, duration) -> times, values

Monitor some property of GPU(s), as specified through the function f, over a given time period, as specified by duration (in seconds). Returns the (relative) times as a Vector{Float64} and the temperatures as a Vector{Vector{Float64}}.

The function f will be called on a vector of devices (CuDevices or NVML.Devices) and should return a vector of Float64 values.

Keyword arguments:

freq (default: 1): polling rate in Hz.

devices (default: NVML.devices()): CuDevices or NVML.Devices to consider.

plot (default: false): Create a unicode plot after the monitoring.

liveplot (default: false): Create and update a unicode plot during the monitoring. Use optional ylims to specify fixed y limits.

title (default: ""): Title used in unicode plots.

ylabel (default: "Values"): y label used in unicode plots.

See: monitor_temperature, monitor_powerusage

PC2GPUBenchmarks.monitor_temperature — Method

monitor_temperature(duration) -> times, temperatures

Monitor the temperature of GPU(s) over a given time period, as specified by duration (in seconds). Returns the (relative) times as a Vector{Float64} and the temperatures as a Vector{Vector{Float64}}.

For general keyword arguments, see monitor_something.