Worker Utilities

Index

References

PC2GPUBenchmarks.@worker — Macro

@worker pid ex

Spawns the given command on the given worker process.

Examples:

@worker 3 PC2GPUBenchmarks.functional()
@worker 3 stresstest(CUDA.devices(); duration=10, verbose=false)

PC2GPUBenchmarks.@worker — Macro

@worker ex

Creates a worker process, spawns the given command on it, and kills the worker process once the command has finished execution.

Implementation: a Julia thread (we use @spawn) will be used to wait on the task and kill the worker.

Examples:

@worker PC2GPUBenchmarks.functional()
@worker stresstest(CUDA.devices(); duration=10, verbose=false)

PC2GPUBenchmarks.@worker_create — Macro

@worker_create n -> pids

Create n workers (i.e. separate Julia processes) and execute using PC2GPUBenchmarks, CUDA on all of them. Returns the pids of the created workers.

PC2GPUBenchmarks.@worker_killall — Macro

Kills all Julia workers.