Utility

Index

PC2GPUBenchmarks.alloc_mem

PC2GPUBenchmarks.clear_all_gpus_memory

PC2GPUBenchmarks.clear_gpu_memory

PC2GPUBenchmarks.functional

PC2GPUBenchmarks.tensorcores

References

PC2GPUBenchmarks.alloc_mem — Method

alloc_mem(memsize::UnitPrefixedBytes; devs=(CUDA.device(),), dtype=Float32)

Allocates memory on the devices whose IDs are provided via devs. Returns a vector of memory handles (i.e. CuArrays).

Examples:

alloc_mem(MiB(1024)) # allocate on the currently active device
alloc_mem(B(40_000_000); devs=(0,1)) # allocate on GPU0 and GPU1

PC2GPUBenchmarks.clear_all_gpus_memory — Function

Reclaim the unused memory of all available GPUs.

PC2GPUBenchmarks.clear_gpu_memory — Function

Reclaim the unused memory of the currently active GPU (i.e. device()).

PC2GPUBenchmarks.functional — Function

Check if CUDA/GPU is available and functional. If not, print some (hopefully useful) debug information.

PC2GPUBenchmarks.tensorcores — Function

tensorcores([enable::Bool; verbose=true])

Switches the CUDA.math_mode between CUDA.FAST_MATH (enable=true) and CUDA.DEFAULT_MATH (enable=false). For matmuls of CuArray{Float32}s, this should have the effect of using/enabling and not using/disabling tensor cores. Of course, this only works on supported devices and CUDA versions.

If no arguments are provided, this functions toggles between the two math modes.

PC2GPUBenchmarks.@unroll — Macro

@unroll N expr

Takes a for loop as expr and informs the LLVM unroller to unroll it N times, if it is safe to do so.

PC2GPUBenchmarks.@unroll — Macro

@unroll expr Takes a for loop as expr and informs the LLVM unroller to fully unroll it, if it is safe to do so and the loop count is known.